nnUNet_WMSegmentation/requirements.txt
# Core framework
nnunetv2

# Data preparation
Pillow
numpy
scikit-learn

# Evaluation
scipy
matplotlib
